mid-way thru refactoring providers
compiling! templates still in progress
