+++ cli/src/providers/gloo/mod.rs
@@ -1,14 +1,15 @@
 use std::rc::Rc;
 use std::sync::Arc;
 
+use clap::crate_version;
 use handlebars::{Handlebars, to_json};
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
 
-use crate::providers::{BoxedCastable, Options, Provider, ProviderError};
+use crate::providers::{Options, Provider, ProviderError};
 use crate::tools::glooctl::GlooCtl;
 use crate::tools::kubectl::KubeCtl;
-use crate::transpiler::{Castable, CastError, ContentType};
+use crate::transpiler::{BoxedCastable, Castable, CastError, ContentType};
 use crate::transpiler::context::{Context, Function, Http};
 
 pub struct ApiProvider {
@@ -18,15 +19,65 @@ pub struct ApiProvider {
 
 impl ApiProvider {
     pub fn new() -> Self {
+        let kubectl = KubeCtl::default();
+        let namespaces = kubectl.get_namespaces().unwrap();
+        let items = namespaces.get("items").unwrap().as_array().unwrap();
+        let is_installed = items
+            .iter()
+            .find(|i| {
+                i.get("metadata")
+                    .unwrap()
+                    .as_object()
+                    .unwrap()
+                    .get("labels")
+                    .unwrap()
+                    .as_object()
+                    .unwrap()
+                    .get("kubernetes.io/metadata.name")
+                    .unwrap()
+                    .as_str()
+                    .unwrap()
+                    .contains("asml-gloo-")
+            })
+            .is_some();
+        // self.gloo_installed = Some(is_installed);
         Self {
             gloo_installed: None,
             options: Arc::new(Options::new()),
         }
     }
+
+    // FIXME not needed?
+    fn cast_kube(&self, ctx: Rc<Context>) -> Result<Vec<String>, CastError> {
+        let mut content = format!("# Generated by assemblylift-cli {}\r\n", crate_version!());
+
+        // we have only one api provider under k8s and it creates a single resource at the 'root'
+        // so we skip looking up the provider and/or making it a side-effect of the root provider
+        // i.e. right now if you're using HTTP with the k8s provider you're using Gloo
+
+        let project_name = ctx.project.name.clone();
+        let has_api = ctx
+            .functions
+            .iter()
+            .filter_map(|f| f.http.as_ref())
+            .collect::<Vec<&Http>>()
+            .len()
+            > 0;
+
+        if has_api {
+            let mut api_provider = ApiProvider::new();
+            content.push_str(
+                &*api_provider.cast(ctx.clone(), None).unwrap()[0] // FIXME unchecked index
+                    .clone(),
+            );
+        }
+
+        Ok(vec![content])
+    }
 }
 
 impl Castable for ApiProvider {
-    fn cast(&mut self, ctx: Rc<Context>, name: &str) -> Result<Vec<String>, CastError> {
+    fn cast(&self, ctx: Rc<Context>, selector: Option<&str>) -> Result<Vec<String>, CastError> {
         let project_name = ctx.project.name.clone();
         {
             // FIXME this should really happen at the beginning of the `bind` command,
@@ -76,33 +127,6 @@ impl Castable for ApiProvider {
 impl Provider for ApiProvider {
     fn name(&self) -> String {
         String::from("k8s-gloo")
-    }
-
-    fn init(&mut self, ctx: Rc<Context>, name: &str) -> Result<(), ProviderError> {
-        let kubectl = KubeCtl::default();
-        let namespaces = kubectl.get_namespaces().unwrap();
-        let items = namespaces.get("items").unwrap().as_array().unwrap();
-        let is_installed = items
-            .iter()
-            .find(|i| {
-                i.get("metadata")
-                    .unwrap()
-                    .as_object()
-                    .unwrap()
-                    .get("labels")
-                    .unwrap()
-                    .as_object()
-                    .unwrap()
-                    .get("kubernetes.io/metadata.name")
-                    .unwrap()
-                    .as_str()
-                    .unwrap()
-                    .contains("asml-gloo-")
-            })
-            .is_some();
-        self.gloo_installed = Some(is_installed);
-
-        Ok(())
     }
 
     fn options(&self) -> Arc<Options> {
@@ -173,6 +197,7 @@ struct RouteData {
     to_function_name: String,
 }
 
+// TODO add another domain to the list for the user's public domain
 static VIRTUALSERVICE_TEMPLATE: &str = r#"apiVersion: gateway.solo.io/v1
 kind: VirtualService
 metadata:
@@ -181,7 +206,7 @@ metadata:
 spec:
   virtualHost:
     domains:
-    - '*'
+    - '{{project_name}}.asml.local'
     {{#if has_routes}}routes:
     {{#each routes}}- matchers:
       - exact: {{this.path}}
