+++ cli/src/transpiler/context.rs
@@ -1,8 +1,15 @@
+use std::fs;
+use std::path::PathBuf;
 use std::rc::Rc;
 use std::sync::Arc;
 
+use clap::crate_version;
+use handlebars::{Handlebars, to_json};
+use serde::Serialize;
+
 use crate::projectfs::Project as ProjectFs;
-use crate::transpiler::{StringMap, toml};
+use crate::providers::SERVICE_PROVIDERS;
+use crate::transpiler::{Castable, CastError, ContentType, context, StringMap, Template, toml};
 
 pub struct Context {
     pub project: Project,
@@ -132,6 +139,54 @@ impl Context {
     }
 }
 
+impl Castable for Context {
+    fn cast(&self, ctx: Rc<Context>, selector: Option<&str>) -> Result<Vec<String>, CastError> {
+        let mut hcl_content = format!("# Generated by assemblylift-cli {}\r\n", crate_version!());
+
+        let mut usermod_path = PathBuf::from(
+            crate::projectfs::locate_asml_manifest()
+                .expect("could not locate assemblylift.toml")
+                .1,
+        );
+        usermod_path.pop();
+        usermod_path.push("user_tf/");
+        let user_inject: bool = fs::metadata(usermod_path.clone()).is_ok();
+        let (remote_state, state_bucket_name, lock_table_name) = match &self.terraform {
+            Some(tf) => (
+                true,
+                Some(tf.state_bucket_name.clone()),
+                Some(tf.lock_table_name.clone()),
+            ),
+            None => (false, None, None),
+        };
+
+        let tmpl = ContextTemplate {
+            project_name: self.project.name.clone(),
+            project_path: self.project.path.clone(),
+            user_inject,
+            remote_state,
+            state_bucket_name,
+            lock_table_name,
+        };
+        hcl_content.push_str(&*tmpl.render());
+
+        for service in &self.services {
+            hcl_content.push_str(
+                &service
+                    .cast(ctx.clone(), Some(&*service.name))
+                    .unwrap()[0]
+                    .clone(),
+            );
+        }
+
+        Ok(vec![hcl_content])
+    }
+
+    fn content_type(&self) -> Vec<ContentType> {
+        todo!()
+    }
+}
+
 pub struct Project {
     pub name: String,
     pub path: String,
@@ -152,6 +207,40 @@ pub struct Service {
 impl Service {
     pub fn option(&self, name: &str) -> Option<&String> {
         self.provider.options.get(name)
+    }
+}
+
+impl Castable for Service {
+    fn cast(&self, ctx: Rc<Context>, selector: Option<&str>) -> Result<Vec<String>, CastError> {
+        let mut content = format!("# Begin service `{}`\n", &self.name);
+
+        let provider_name = self.provider.name.clone();
+        let mut service_provider = SERVICE_PROVIDERS
+            .get(&provider_name)
+            .expect(&format!(
+                "could not find service provider named {}",
+                provider_name
+            ))
+            .lock()
+            .unwrap();
+
+        service_provider
+            .set_options(self.provider.options.clone())
+            .expect("unable to set provider options");
+        // service_provider
+        //     .init(ctx.clone(), self.name.into())
+        //     .expect("unable to initialize service provider");
+
+        let service_artifact = service_provider
+            .cast(ctx.clone(), selector)
+            .expect("unexpected error transforming service");
+        content.push_str(&service_artifact[0].clone()); // FIXME unchecked index
+
+        Ok(vec![content])
+    }
+
+    fn content_type(&self) -> Vec<ContentType> {
+        todo!()
     }
 }
 
@@ -197,3 +286,61 @@ pub struct Iomod {
     pub coordinates: String,
     pub version: String,
 }
+
+#[derive(Serialize)]
+pub struct ContextTemplate {
+    pub project_name: String,
+    pub project_path: String,
+    pub user_inject: bool,
+    pub remote_state: bool,
+    pub state_bucket_name: Option<String>,
+    pub lock_table_name: Option<String>,
+}
+
+impl Template for ContextTemplate {
+    fn render(&self) -> String {
+        // to_json(&self);
+        let mut reg = Box::new(Handlebars::new());
+        reg.register_template_string("hcl_template", Self::tmpl())
+            .unwrap();
+        reg.render("hcl_template", &self).unwrap()
+    }
+
+    fn tmpl() -> &'static str {
+        r#"terraform {
+    required_providers {
+        docker = {
+            source  = "kreuzwerker/docker"
+            version = ">= 2.16.0"
+        }
+        kubernetes = {
+          source  = "hashicorp/kubernetes"
+          version = ">= 2.0.0"
+        }
+    }
+}
+
+locals {
+    project_name = "{{project_name}}"
+    project_path = "{{project_path}}"
+}
+{{#if user_inject}}
+module "usermod" {
+  source = "../user_tf"
+}
+{{/if}}
+{{#if remote_state}}
+terraform {
+  backend "s3" {
+    encrypt = true
+    bucket = "{{state_bucket_name}}"
+    dynamodb_table = "{{lock_table_name}}"
+    key    = "terraform.tfstate"
+    region = "us-east-1"
+  }
+}
+{{/if}}
+"#
+    }
+}
+
